Adds 'import project from repository' feature Project tab.
Modifies the Project tab to include an 'import from repository' option.
Includes an ImportTemplateWizard, similar to the ProjectUploadWizard.
Templates are pre-built projects that can include 0 or more
assets, components, and/or blocks to serve as 'scaffolding'
for lessons and tutorials.

Template repositories may be stored in the build/war directory on
the AppEngine server or may be on any reachable web serer. The
repository must be structured in a particular way and use certain
naming conventions.  These rules are described in TemplatesUploadWizard.java.

If templates are stored on a static website. Your webserver must
implement "Cross-origin Resource Sharing (CORS). With Apache you can
do this by using mod_headers placing a .htaccess file in the templates
directory that contains the one line:

  'Header set Access-Control-Allow-Origin "*"'.

External repositories can be added/removed by the user at runtime.

Change-Id: I3b0e8b57d70a9e1b80ed7ce5d6f21248fa130213