First Release of Emulator Compatible NewBlocks Companion.
Adapt Companion App to recognize when it is running on the emulator
and not interact with the user or the rendezvous server. Just start
listening. Use a "standard" HmacKey (because security is not an issue
here).

Add "Connect to Emulator" button that connects the blocks to the
Companion App running in the emulator.

Change-Id: Ia63eabacdc3339cf4d31505d8f6ba9bcd6425a66