Revises Texting component so it works with standard loops (issue #478)
Sending message in a loop should work both in the packaged app and
in the REPL.

If GV is enabled when an app starts, Oauth authentication is performed
on start up and subsequent attempts to send messages proceed without
additional authentication.

If GV is disabled, messages are sent through built-in Sms.

If GV is enabled but the authToken is null, messages are queued while
authentication is performed.

NOTE: To test the receiving-in-background feature of the Texting component,
you must package the app.

Change-Id: I535c2da3063468d13f6c82a1eba1d551d797532c