Use Base64 to encode binaries (assets) from Server to Blocks.
GWT RPC was serializing binary objects in a very inefficient
way (read: making a JSON array of bytes of the object, expanding its
size by a factor of between 8 and 10!!!). We now use base64 to encode
binaries.

Also commented out the logging of GWT RPC Responses. It was just
creating clutter.

Change-Id: Ic76688f401a12de0ad0c6ef541ce8bfb04ab893d