Automatic Updating of Companion
The Companion will now auto-update itself unless it was downloaded
from the Play Store. Auto-Updating is a violation of the Google Play
Store developer policies. However side loaded Companions will
auto-update. This is of particular use in the Companion installed
within the setup tools. This will help us to avoid having to update
the Setup Tools with each change to the components, which requires a
Companion update.

Change-Id: I92069d38f8a838d4d44ca9382ea03304585824b4