Fix issue #668 by adding red error triangles to duplicate component event handlers.
This version uses ":" rather than "_" as an internal separator
character to defend against problems involving names with underscores.

It also abstracts over error strings (for internationalization)
in blocklyeditor/src/warningHandler.js.

This version abandons the earlier notion of special treatment of
singleton components = components (like AccelerometerSensor) for
which there can be conceptually only one instance. In this version,
all component event handlers are treated the same way.

Care was taken to avoid quadratic-time determination of duplicates
that was a problem in an earlier experimental solution to this problem
and was unacceptable for screens with large numbers of component event
handlers. In this new submission, duplicates event handlers are found
in a process that is linear in the number of component event handler
blocks, and each such block runs a constant-time error check to add
(or remove) the red error triangle.

Change-Id: Id7b0a69a4faf9028b5f534de515201d5a680ad2a