Alter the way Companion Updates Happen – Use Push
The AI Client code (in the browser) now downloads the new
Companion (from the service) and “pushes” it to the Companion using a
mechanism similar to the mechanism used to send assets to the
Companion. This eliminates difficulties encountered when a proxy
exists between the Internet and the AI user’s location.

Change-Id: I5d91de137d2ed57199949b0761fdf13fa24fa8fe