Increase the reliability of starting the Emulator
Increase the “ubergiveup” threshold to 8. Add the option for the user
to tell App Inventor to keep trying even after the ubergiveup
interval. Add an extra intent to start the Companion if it hasn’t
shown a response yet. This should increase the chances of the
Companion starting on very slow computers. Changed some timers
as well.

Minor change to the build environment so that versions are marked
“dirty” (in the Help->About Menu) if the system was built from a work
tree that has changes that are not yet committed in any branch.

Change-Id: I7bd42d36b3b005c2e072aad60b65afe555e8f2e1