Permit People to Continue Past Double Session Warning.
Add a third option to the dialog box to permit people to continue
working even though there is a second account logged in. This will
permit classes that use one ID to continue to work. It should be OK as
long as two people do not work on the same project.

Change-Id: Iaaeb7afeea36cc0820cb12f75a282f6d8cccdec5