Add Rendezvous Server settings to service.
Permit location of the rendezvous service to be specified in system
properties (which can be set in appengine-web.xml for App Engine
version).

Create new system “Config” record that is fetched at Ode start
time via the userInfoProvider.getSystemConfig() call. This call
provides both the User object as well as system configuration. For now
this includes the location of the Rendezvous server. Note: This call
replaces the call to getUserInformation().

Change-Id: Ie95fca453288c2d02223d6b99ce3879301a870eb