Arrange for Companion to report its package name.
Arrange for Companion to report its package name, this permits the
service to determine if the correct companion is being used. This is
of particular value to people compiling their own version of App
Inventor and want to make sure their end-users are using the
corresponding Companion. The code to check the package name in the
service is not yet written, this change is just on the Companion side
to report its package name along with its version.

Change-Id: If1b9624ab60830b336bac8122eb8dbb9fb34a167