Quick Fix to the Truncated Blocks Detection Code
Updated the length of the “empty” workspace check because the length
of the empty workspace became larger (YaVersion added a digit) and the
check was failing to correctly detect an empty workspace. Added some
extra digits as well as a pad. A future update will measure the actual
length of an empty workspace for the comparison.

Change-Id: I70ee36083a7568ebfa3ea405eb9e8817dea561b6