Ensure that Built APK files are valid
If we fail to fetch a file from the Google Cloud Store, we substitute
a zero length file. This change alters this behavior to cause a fatal
error if this happens during an attempt to build a project. This
ensures that zero length assets do not wind up in a built APK file.

Change-Id: I09f841f06dacb413adb49056c73b62f65cddcaa9