Add the Deobsfucate Text block.
This block works like a text block. However the code it generates is
obsfucated so that the text string itself is obscured from view in a
packaged application. It is a good block to use to store an API key or
other information that is sensitive.

Note: This is not a perfect solution. A determined person can still
take a packaged file and extract the text string. However it takes
some effort.

Change-Id: I481e890b2230f69877b61a697743a921fb15fb83