Update How App Inventor is served and cached
Improve caching behavior so that when a release is made updates are
more likely to be seen by end-users faster with less inconsistency
between the Designer and the Blocks Editor. This change also arranges
for new versions of Blockly to be seen on each build, hopefully
eliminating the need to disable caching during development.

Change-Id: Ie6e38c46db5fd4208fcc4bb2c680ebefdccc1061