Properly Report Errors during initial form load
When we load the initial form, we tend to issue many Yail evaluation
requests in a row. Low level code in replmgr.js then consolidates
these changes into one XHR call instead of sending each one in its own
request. This improves form loading performance significantly.

However if an error occurs during code that is consolidated, we cannot
determine where the error occurred. Also, because the code is
consolidated into one (begin ...) block, an error in the middle causes
later s-expressions to not be evaluated. This leaves the Companion in
an undefined and usually broken state.

This code change detects an error when s-expressions are consolidated
and arranges to resend the form (reset the screen) without the
optimization in place. This results in the older “slow” code being
used, but better handles errors in the consolidated code. After 20
successful calls sending s-expressions over, the optimized code is
re-enabled.

Change-Id: Ic29ecdca7d3842f91829169842802419dd1a9fcc