Remove the “salvageAll” feature from the Gallery
The “salvageAll” feature, available only to Gallery moderators was
useful back when the Gallery only had a few apps. However there are
too many apps today for it to be of much use. It would just use a lot
of resources and then fail. Because the salvageAll link is very near
the “salvage” and other related links, there is a risk of accidentally
invoking it. So we remove it here.

Change-Id: I86686789e224b51307ca684aa2d9d1cd3a4d7a0c