“Debounce” the Project Start Button
Disable the Start New Project button when it is clicked. We do this
because on slow machines people click it multiple times while the
start new project wizard is starting. This then causes a second wizard
to start and a very confused user experience.  We will enable the
button again when we re-visit the Project List page

Change-Id: Ie060ae26d6c4f6dc44c7b4937a3e8a16b4edaf3b