MIT App Inventor Extensions
Add the ability to add non-visible components, written in Java, at
run-time. Components are built using the MIT App Inventor sources and
then packaged into .aix files. These files can then be loaded into a
project which will extend App Inventor’s functionality.

Note: Projects with extensions cannot currently be uploaded to the
Gallery due to legal concerns about placing binary code in the
Gallery (the binary of the extension is carried in the source of the
project).

This commit represents the work of several people over almost a year’s
time. They are (alphabetic order):

Author: E-Hon <e.hon.dev@gmail.com>
Author: Jeffrey I. Schiller <jis@mit.edu>
Author: Justus <quixoticquixer@gmail.com>
Author: Mouhamadou Oumar Sall <mouha.oumar@gmail.com>
Author: William S. Byrne <will2596@gmail.com>

Change-Id: I90d3cb9f5485770f34ce577a0b58a587799fca32