Updates to Firebase
Add “ClearTag” to FirebaseDB as well as a way to get a list of known
tags for a project. This helps facilitate cleaning up the tags used by a
project.

Add Persist Designer Property to FirebaseDB. When set to true, variables
will be saved when an App exits, even if there is no network
connectivity

Change-Id: I37c27c889693f441ee2a7ad366b754d5c7e5ddac